Split out key tests between windows & linux since the behavior is now different
