+++ pkg/network/event_windows_test.go
@@ -8,6 +8,11 @@ package network
 import (
 	"testing"
 
+	"golang.org/x/exp/slices"
+
+	"github.com/DataDog/datadog-agent/pkg/network/types"
+	"github.com/DataDog/datadog-agent/pkg/process/util"
+
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 )
@@ -39,4 +44,29 @@ func TestNetshParse(t *testing.T) {
 		assert.Equal(t, uint16(49152), low)
 		assert.Equal(t, uint16(65535), hi)
 	})
+}
+
+func TestKeyTuplesFromConn(t *testing.T) {
+	sourceAddress := util.AddressFromString("1.2.3.4")
+	sourcePort := uint16(1234)
+	destinationAddress := util.AddressFromString("5.6.7.8")
+	destinationPort := uint16(5678)
+
+	connectionStats := ConnectionStats{
+		Source: sourceAddress,
+		SPort:  sourcePort,
+		Dest:   destinationAddress,
+		DPort:  destinationPort,
+	}
+	keyTuples := ConnectionKeysFromConnectionStats(connectionStats)
+
+	assert.Len(t, keyTuples, 1, "Expected different number of key tuples")
+	assert.True(t, slices.ContainsFunc(keyTuples, func(keyTuple types.ConnectionKey) bool {
+		sourceAddressLow, sourceAddressHigh := util.ToLowHigh(sourceAddress)
+		destinationAddressLow, destinationAddressHigh := util.ToLowHigh(destinationAddress)
+		return (keyTuple.SrcIPLow == sourceAddressLow) && (keyTuple.SrcIPHigh == sourceAddressHigh) &&
+			(keyTuple.DstIPLow == destinationAddressLow) && (keyTuple.DstIPHigh == destinationAddressHigh) &&
+			(keyTuple.SrcPort == sourcePort) && (keyTuple.DstPort == destinationPort)
+	}), "Missing original connection")
+
 }
